fix: [SMB client] Incorrect openFile() attributes
* Ref: https://github.com/hierynomus/smbj/issues/786

+++ source/core/data/src/main/kotlin/com/xayah/core/data/repository/PackageRepository.kt
@@ -511,7 +511,7 @@ class PackageRepository @Inject constructor(
                     val remote = entity.remote
                     val remoteArchivesPackagesDir = pathUtil.getCloudRemoteAppsDir(remote)
                     val src = "${remoteArchivesPackagesDir}/${p.archivesRelativeDir}"
-                    client.deleteRecursively(src)
+                    if (client.exists(src)) client.deleteRecursively(src)
                 }
             }.onFailure(rootService.onFailure).isSuccess
         }
